pyproject.toml
[tool.poetry]
name = "translatorai"
version = "0.1.0"
description = ""
authors = ["Duc Do <ducdh1210@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
openai = "^1.33.0"
langchain = "^0.2.3"
python-dotenv = "^1.0.1"
langchain-openai = "^0.1.8"
pypdf = "^4.2.0"
langgraph = "^0.0.66"
pygraphviz = "^1.13"
fastapi = "^0.111.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
